feat: moved menu options to separate files
