+++ app/cdap/components/GridTable/components/NestedMenu/menuOptions/datatypeOptions.ts
@@ -1,0 +1,63 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+export const DATATYPE_OPTIONS = [
+  {
+    value: 'string',
+    label: 'string',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'boolean',
+    label: 'boolean',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'integer',
+    label: 'integer',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'long',
+    label: 'long',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'short',
+    label: 'short',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'float',
+    label: 'float',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'double',
+    label: 'double',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'decimal',
+    label: 'decimal',
+    supported_dataType: ['all'],
+  },
+  {
+    value: 'bytes',
+    label: 'bytes',
+    supported_dataType: ['all'],
+  },
+];
