+++ app/cdap/components/GridTable/index.tsx
@@ -27,7 +27,7 @@ import React, { useEffect, useState } from 'react';
 import { useLocation, useParams } from 'react-router';
 import { flatMap } from 'rxjs/operators';
 import { objectQuery } from 'services/helpers';
-import ToolBarList from './components/TransfomationToolbar';
+import ToolBarList from 'components/GridTable/components/TransfomationToolbar';
 import BreadCrumb from './components/Breadcrumb';
 import GridHeaderCell from './components/GridHeaderCell';
 import GridKPICell from './components/GridKPICell';
